refactor integration server utils to separate file so it can be used in other tests add web3 position getters
